requirements.txt
gunicorn
django
psycopg2-binary
Pillow
djangorestframework
djangorestframework-simplejwt
requests
django-extra-fields